prototyping go rules implementation
GoRulesValidator reads a directory of .yml files that contain sparql and
rule metadata.  This gets integrated into the MolecularModelManager
stack.  Right now, when a request for the gpad for a model is received,
it will run all the sparql in the rules and make a little report. Need
to (1) figure out how the rules will be called (e.g. what parameters,
when) (2) decide what to send back and (3) refine the overall pattern
for what constitutes a rule.  Now, we can't use any sparql returns aside
from a count because the same result parser is applied to every sparql
rule request.