Added shex support to GoRules validator framework
No useful rules yet, just the skeleton that executes them.  shex rules
currently need to be in separate files from the rule metadata.  the rule
file contains a path to the shex file.  (could be made a URL or
something more clever).  Probably some more work with yaml and the shex
parser would allow them to be embedded like sparql.

Next steps: define and implement an API for the validation results,
write the rules, add to Noctua UI.