reverting back to owlapi4.2.8
Version 4.5.7 causes MolecularModelManagerTest.testExportImport() to fail in a bizarre and troublesome way.  It will hang forever as the MolecularModelManager attempts to save the abox ontology to bytestream (about line 818).

I reproduced the problem in separate, simple test file.  It only happens if the OntologyManager.applyChange method is called before attempting to write to the stream.  e.g. like this:
OWLOntologyID ontologyID = new OWLOntologyID(Optional.of(expectedABoxIRI), Optional.of(expectedABoxIRI));
				manager.applyChange(new SetOntologyID(aBox, ontologyID));
if that change is not applied it will run just fine.
Yay..  So rolling back to working version.