scratchily working cut at #254
Also does shex validation, following exact pattern used in minerva service now.  Should clean up the output and probably get rid of the shex-specific method (and add some of the customization parameters in that one into this one).
Note that unless you use a catalog to deny its attempts to load the full go-lego with all of neo, this will require a lot of memory to run.  (Out of mem with Xmx12g).  But of course if you don't have the little neo bits that are used in your models, it won't work either.