OWLTools is no longer a dependency
I believe everything is working.  All tests are passing. Server runs successfully locally.  Command line calls are all working the same way as before.  (With some minor changes to the WIP validation call.)

There are a few files copied over from OWLTools into minerva-core.  If we care we can dig through them and eliminate more cruft as we like.  They are easily found in the owltools packages in that project.