fixing confusion with travis mode
travis_mode is used only for go_shapes to fail when something is wrong.  This change allows the shex validation to run when the OWL validation succeeds - and when we are expecting a failure.  if we want to separate owl and shex validation and use this kind of set up we would want to introduce another yet another parameter to indicate travis-owl travis-shex travis-both ...   currently equates to travis-shex