reasonable first cut at testing framework for search
A lot more can be done here, but this gets us started.  Could do: more specific tests (e.g. ordering checks), combinatorial parameter sweep tests, etc. etc.  As it stands we can check that the handler is launchable and that all of the basic parts of the API are operating.