investigating slow server performance
Identified the conversion of the OWL model into the json structure as by far the slowest part of the whole thing.  Tricky recursive function to blame.  Adding the reasoner appears to slow things down a lot but in fact, its just because it gives the renderer more work to do.
no reasoning:
2020-03-16 16:56:44,438 INFO  (JsonOrJsonpBatchHandler:218) model rendering is slowing things down...
2020-03-16 16:56:49,175 INFO  (JsonOrJsonpBatchHandler:220) See...

with reasoning
2020-03-16 16:57:00,648 INFO  (JsonOrJsonpBatchHandler:218) model rendering is slowing things down...
2020-03-16 16:57:48,679 INFO  (JsonOrJsonpBatchHandler:220) See...