added 'expand' as a non-default option for #291
now need to add the parameter expand to a term search query to get it to include results from matches on the subclass closure.  The value of the parameter does not matter and can be empty.  e.g. http://127.0.0.1:6800/search/?term=http%3A%2F%2Fpurl.obolibrary.org%2Fobo%2FGO_0140312&expand=