refactor taxon handling for #290
This includes a method to update all models in a given journal with taxon information as metadata on the models.  It will need to be run on the input database for the taxa-related search features to work.  Models saved using this build will include the taxon data.  Notable performance improvements over last incarnation.

Batch update all taxon metadata.
minerva-cli.sh
--add-taxon-metadata
-j blazegraph-gocam-db.jnl
-ontojournal /tmp/blazegraph.jnl