this seems to be the main problem
Somehow..  I think the re-use of the go-cam's IRI as the IRI for the schema ontology confused the system.  Following a GPAD export, the go-cam model had nothing in it except for all of the object property assertions from the original imported go-lego ontology.  Taking that out stabilized my system locally.