adding a shex validator test into minerva-core
this is mostly redundant with a test in minerva-server .  Added here because it makes more sense to be here.  Currently the validation test in minerva-server is more complete (includes Arachne inference step while this does not) and more exactly matches the conditions in the server (running with models in a MMManager).  This one is useful for testing as it runs faster.