added hacky method to try to cure ontology loader problems
fixBrokenObjectPropertiesAndAxioms alleviates problems with object properties being read in as annotation properties.  It uses the RO namespace to guess what should be an object property and what should not.  With the OWL-API update, this may not be useful anymore.  Also added a simple test to see if a gocam has the go-lego import statement.